Move Message to entities
