+++ ayatori/src/execution.rs
@@ -1,5 +1,4 @@
 mod evidence;
-mod message;
 mod session;
 mod session_id;
 mod storage;
@@ -8,7 +7,6 @@ mod task;
 pub(crate) use session::SessionData;
 
 pub use evidence::{Evidence, EvidenceError};
-pub use message::Message;
 pub use session::{Session, SessionReport, TaskError};
 pub use session_id::SessionId;
 pub use task::Task;
